requirements.txt
google-generativeai
fastapi
requests 
beautifulsoup4